rename: ValidationUtils 클래스를 common/util로 이동
